# api/openapi.yaml
openapi: 3.0.3
info:
  title: FolioPath HTTP API
  version: 0.1.0
  description: |
    Authoritative HTTP contract through the FolioPath POST-MVP-5 scope,
    revision 2. Revision-1 image semantic operations are contract-frozen.
    Revision-2 controlled tag suggestions, video semantic search, and local
    anonymous face/person curation remain gated from production composition
    until INT-S1R2 is accepted.

    All business endpoints use an authenticated same-origin browser session.
    State-changing authenticated operations additionally require the
    session-bound `X-CSRF-Token`. First-run setup and login do not yet have a
    session, so they require a validated same-origin `Origin` header.

    IDs and cursors are opaque. Media content endpoints accept only resource
    IDs, never filesystem paths. Management responses may expose a library
    root relative to the allowed `/library` container mount and a server-built
    `/library/...` display label; no response may expose a host path.
  license:
    name: AGPL-3.0-or-later
    url: https://www.gnu.org/licenses/agpl-3.0.html
...
servers:
  - url: /
    description: Same-origin FolioPath server
security:
  - cookieAuth: []
...
paths:
...
  /api/v1/catalog/state:
    get:
      tags: [Catalog]
      operationId: getCatalogState
      summary: Get the lightweight catalog content revision
      description: |
        Returns a monotonic notification revision for authenticated clients.
        It advances after a reliable full generation is published and after
        each committed automatic-discovery reconciliation batch. It is not a
        browse/search cursor revision and does not authorize filesystem
        access. A matching If-None-Match returns 304 without a body.
      x-requirements: [FR-SCN-014, NFR-PERF-003, NFR-PRIV-001]
      x-error-codes:
        '401': [authentication_required, session_expired]
        '429': [rate_limited]
        '500': [internal_error]
      parameters:
        - $ref: '#/components/parameters/IfNoneMatchHeader'
      responses:
        '200':
          description: Current catalog content revision.
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            Cache-Control:
              $ref: '#/components/headers/NoStore'
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogState'
        '304':
          description: Catalog content has not changed for the supplied validator.
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            Cache-Control:
              $ref: '#/components/headers/NoStore'
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/v1/libraries/{libraryId}/directories:
    get:
      tags: [Catalog]
      operationId: listDirectories
      summary: List direct child directories from the reliable index
      description: |
        Omit `parentId` for the logical library root. Supplying that library's
        indexed root-directory ID is equivalent and is normalized to the same
        cursor scope. Every direct child in the preserved index is returned,
        including directories with no supported media. A parent from another
        library is indistinguishable from a missing directory and returns
        `directory_not_found`.

        Optional `q` filters every direct child in the reliable index, not just
        the current page. The server trims outer Unicode whitespace, applies
        Unicode NFKC plus full case folding, splits on Unicode whitespace,
        removes duplicate terms, and requires every term to occur as a literal
        substring of the normalized direct-child name. Diacritics are
        preserved; one- and two-character terms are valid; punctuation and
        FTS syntax have no operator meaning.

        Results use `(naturalNameKey ASC, name ASC, id ASC)` keyset order.
        Cursors bind the library, normalized parent, normalized query terms,
        search profile, ordering version, and reliable catalog generation.
        Changing or clearing `q`, or advancing the reliable generation, makes
        an old cursor fail with `invalid_cursor`. An offline library still
        returns its preserved index with `200`; pending/scanning/offline state
        is read from the Library resource, so an empty page alone is never
        proof that the filesystem is empty. The request never traverses the
        filesystem.
      x-requirements: [FR-BRW-001, FR-BRW-005, FR-BRW-008, FR-BRW-009, FR-BRW-010]
      x-error-codes:
        '400': [invalid_request, invalid_cursor]
        '401': [authentication_required, session_expired]
        '404': [library_not_found, directory_not_found]
        '429': [rate_limited]
        '500': [internal_error]
      parameters:
        - $ref: '#/components/parameters/LibraryIDParameter'
        - $ref: '#/components/parameters/ParentDirectoryIDParameter'
        - $ref: '#/components/parameters/DirectorySearchQueryParameter'
        - $ref: '#/components/parameters/CursorParameter'
        - $ref: '#/components/parameters/LimitParameter'
      responses:
        '200':
          description: Direct children in natural-name ascending order with stable ID tie-breaker.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DirectoryPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/directories/{directoryId}:
    get:
      tags: [Catalog]
      operationId: getDirectory
      summary: Get an indexed directory and safe breadcrumbs
      description: |
        The indexed root directory is a valid resource: its public `name` is
        the current media-library display name, `relativePath` is empty, and
        `parentId` is null. `breadcrumbs` is ordered root-to-current and always
        includes both endpoints, so the root detail contains exactly one item.
        Breadcrumbs expose only IDs, display names, and library-relative paths.
        Offline libraries return preserved indexed detail with `200`; this
        operation never checks or traverses the filesystem.
      x-requirements: [FR-BRW-001, FR-BRW-003, FR-BRW-008, NFR-PRIV-001]
      x-error-codes:
        '401': [authentication_required, session_expired]
        '404': [directory_not_found]
        '429': [rate_limited]
        '500': [internal_error]
      parameters:
        - $ref: '#/components/parameters/DirectoryIDParameter'
      responses:
        '200':
          description: Directory detail. All paths are relative to its media-library root.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DirectoryDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/libraries/{libraryId}/assets:
    get:
      tags: [Catalog]
      operationId: listLibraryAssets
      summary: Browse or search media within a media library
      description: |
        Omit `directoryId` to use the logical library root. Supplying that
        library's indexed root-directory ID is equivalent and is normalized to
        the same cursor scope. A directory from another library is
        indistinguishable from a missing directory and returns
        `directory_not_found`.

        With no `q`, `recursive=false` returns only assets whose parent is the
        selected directory. `recursive=true` includes the selected directory
        and all indexed descendants; at the root it includes the whole library.

        With `q`, no `directoryId`, and no `recursive` parameter, the scope is
        the entire current library. Supplying `recursive=false` explicitly
        selects only direct assets at the logical library root; supplying
        `recursive=true` selects the root and all descendants. With both `q`
        and `directoryId`, the scope is that current directory; omitted/false
        `recursive` searches direct assets and true includes descendants. This
        explicit recursive parameter lets the browse UI distinguish a filtered
        root directory from the distinct whole-library search scope. Every item
        retains its source `directoryId` and library-relative path.

        Search input is plain text, never FTS syntax. The server trims outer
        Unicode whitespace, applies Unicode NFKC and full case folding, splits
        on Unicode whitespace, removes duplicate terms, and requires every
        term to occur as a literal substring in either the normalized filename
        or normalized library-relative path. Diacritics are preserved. Quotes,
        percent signs, underscores, dashes, dots, and path separators have no
        operator meaning. One- and two-character terms remain valid.

        `kind` and the half-open modification-time interval
        `[modifiedFrom, modifiedBefore)` filter the selected scope. Bounds are
        UTC RFC 3339 instants and `modifiedFrom` must precede
        `modifiedBefore` when both are present. The date field is filesystem
        modification time, never EXIF or container creation time.

        With no search and `recursive=false`, the default order is natural name
        ascending. Recursive or searched results default to file modification
        time descending. Name order groups media by source folder before
        applying natural filename order. Within one library its tuple is
        `(directoryPath, naturalNameKey, name, relativePath, id)`; modified-time order uses
        `(modifiedAt, id)`; size order uses `(sizeBytes, id)`, with every tuple
        component following the requested direction. No relevance sort exists.
        Cursors bind the
        library, canonical scope, effective recursive mode, normalized terms,
        kinds, time bounds, effective sort/order, ordering/profile versions,
        and reliable catalog generation. Any bound-value or generation change
        makes an old cursor fail with `invalid_cursor`; it never falls back to
        page one.

        Offline libraries return preserved indexed assets with `200` and mark
        their source availability. Pending/scanning/offline state comes from
        the Library resource, so an empty page is not by itself a final empty
        state. The request never traverses the filesystem.
      x-requirements: [FR-BRW-002, FR-BRW-003, FR-BRW-005, FR-BRW-009, FR-SRH-001, FR-SRH-002, FR-SRH-004]
      x-error-codes:
        '400': [invalid_request, invalid_cursor]
        '401': [authentication_required, session_expired]
        '404': [library_not_found, directory_not_found]
        '429': [rate_limited]
        '500': [internal_error]
      parameters:
        - $ref: '#/components/parameters/LibraryIDParameter'
        - $ref: '#/components/parameters/DirectoryFilterParameter'
        - $ref: '#/components/parameters/RecursiveParameter'
        - $ref: '#/components/parameters/SearchQueryParameter'
        - $ref: '#/components/parameters/KindFilterParameter'
        - $ref: '#/components/parameters/ModifiedFromParameter'
        - $ref: '#/components/parameters/ModifiedBeforeParameter'
        - $ref: '#/components/parameters/SortParameter'
        - $ref: '#/components/parameters/OrderParameter'
        - $ref: '#/components/parameters/CursorParameter'
        - $ref: '#/components/parameters/LimitParameter'
      responses:
        '200':
          description: Stable indexed media page.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/v1/assets/{assetId}:
    get:
      tags: [Media]
      operationId: getAsset
      summary: Get indexed media details and derived-state status
      x-requirements: [FR-MED-005, FR-MED-006, FR-MED-007]
      parameters:
        - $ref: '#/components/parameters/AssetIDParameter'
      responses:
        '200':
          description: Asset details. `relativePath` is relative to the media-library root.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Asset'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: foliopath_session
      description: |
        High-entropy, opaque session cookie backed by a revocable server-side
        session. It is HttpOnly and SameSite=Strict; deployments using HTTPS
        set Secure. The database stores no replayable plaintext token.
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: |
        Session-bound token returned by setup, login, and session responses.
        It is required together with `cookieAuth` on authenticated
        state-changing operations. It must never be placed in a URL.
  parameters:
...
    LibraryIDParameter:
      name: libraryId
      in: path
      required: true
      description: Opaque media-library ID. It contains no path information.
      schema:
        $ref: '#/components/schemas/ResourceID'
...
    ParentDirectoryIDParameter:
      name: parentId
      in: query
      required: false
      description: Opaque parent directory ID. Omit it for the media-library root.
      schema:
        $ref: '#/components/schemas/ResourceID'
...
    IfNoneMatchHeader:
      name: If-None-Match
      in: header
      required: false
      description: Representation validator from an earlier response.
      schema:
        type: string
        maxLength: 512
...
  headers:
    NoStore:
      description: Authentication state and credentials must not be stored by browsers or intermediaries.
      schema:
        type: string
        enum: [no-store]
    RequestID:
      description: Server-generated opaque correlation ID. Client-supplied values are not trusted.
      schema:
        type: string
        pattern: '^req_[A-Za-z0-9_-]+$'
        maxLength: 128
...
    ETag:
      description: |
        Strong opaque representation validator. For a Library it changes when
        any represented field changes, including scan-derived status/counts.
      schema:
        type: string
        example: '"fp-8b51cdd5"'
...
  responses:
...
    Unauthorized:
      description: A valid administrator session is required or the session has expired.
      headers:
        Cache-Control:
          $ref: '#/components/headers/NoStore'
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    TooManyRequests:
      description: A per-endpoint authentication, browse, scan, thumbnail, or media limit was reached.
      headers:
        Cache-Control:
          $ref: '#/components/headers/NoStore'
        Retry-After:
          $ref: '#/components/headers/RetryAfter'
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: Unexpected failure represented by a safe public code and request ID.
      headers:
        Cache-Control:
          $ref: '#/components/headers/NoStore'
        X-Request-ID:
          $ref: '#/components/headers/RequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    CatalogState:
      type: object
      additionalProperties: false
      required: [contentRevision]
      properties:
        contentRevision:
          type: integer
          format: int64
          minimum: 1
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/Error'
...